+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - `theme.path` option of `code_highlight` plugin.
 
 ### Fixed
+- Removed markdown syntax for some values in `metas`, `feed` and `json_ld`.
 - `relative_urls` plugin with no pretty urls [#711].
 - Updated deps: `sass`, `preact`, `xml`, `katex`, `jsr`, `pagefind`, `highlight.js`, `tailwindcss`, `magic-string`, `std`, `esbuild`, `liquid`, `markdown-it-attrs`, `unocss`, `vento`, `deno_dom`, `unocss`, `satori` and some icons.
 - Several changes on `esbuild` plugin: